moving dataset zoo to datasets/ subpackage
